celt/silk/entdec: drop casts that clippy 1.98 now flags
The nightly toolchain bump pulled clippy 0.1.98, whose `unnecessary_cast`
is stricter: it now flags same-type casts whose type comes from a function
return or a prior cast - e.g. `ec_tell(dec) as i32` (ec_tell already returns
i32, x8 in celt.rs), the trailing `as i32` on two already-i32 bands.rs
expressions, and a doubly-cast `(.. as u32) as u32` in entdec. Also clears one
`explicit_counter_loop` in the SILK PLC: `pred_lag_off` was a manual counter
bumped in lockstep with the loop, now derived from the loop index.

All identity no-ops bar the loop rewrite (obviously equivalent): RFC vectors
stay bit-exact in both float and fixed-point (quality 97.0750 / 97.7667).

+++ src/bands.rs
@@ -598,12 +598,12 @@ pub fn quant_band(
                 let fm = ec_decode(ec, ft);
 
                 if (fm as i32) < (((qn_val >> 1) * ((qn_val >> 1) + 1)) >> 1) {
-                    itheta = (((8 * fm + 1).isqrt() as i32 - 1) >> 1) as i32;
+                    itheta = ((8 * fm + 1).isqrt() as i32 - 1) >> 1;
                     let fs = itheta + 1;
                     let fl = (itheta * (itheta + 1)) >> 1;
                     ec_dec_update(ec, fl as u32, (fl + fs) as u32, ft);
                 } else {
-                    itheta = ((2 * (qn_val + 1) - (8 * (ft - fm - 1) + 1).isqrt() as i32) >> 1) as i32;
+                    itheta = (2 * (qn_val + 1) - (8 * (ft - fm - 1) + 1).isqrt() as i32) >> 1;
                     let fs = qn_val + 1 - itheta;
                     let fl = ft as i32 - (((qn_val + 1 - itheta) * (qn_val + 2 - itheta)) >> 1);
                     ec_dec_update(ec, fl as u32, (fl + fs) as u32, ft);
